ref: extract shared method for package changes
