+++ src/managers/conda/condaPackageManager.ts
@@ -11,7 +11,6 @@ import {
     DidChangePackagesEventArgs,
     IconPath,
     Package,
-    PackageChangeKind,
     PackageManagementOptions,
     PackageManager,
     PythonEnvironment,
@@ -20,18 +19,8 @@ import {
 import { showErrorMessageWithLogs } from '../../common/errors/utils';
 import { CondaStrings } from '../../common/localize';
 import { withProgress } from '../../common/window.apis';
+import { getPackageChanges } from '../common/packageChanges';
 import { getCommonCondaPackagesToInstall, managePackages, refreshPackages } from './condaUtils';
-
-function getChanges(before: Package[], after: Package[]): { kind: PackageChangeKind; pkg: Package }[] {
-    const changes: { kind: PackageChangeKind; pkg: Package }[] = [];
-    before.forEach((pkg) => {
-        changes.push({ kind: PackageChangeKind.remove, pkg });
-    });
-    after.forEach((pkg) => {
-        changes.push({ kind: PackageChangeKind.add, pkg });
-    });
-    return changes;
-}
 
 export class CondaPackageManager implements PackageManager, Disposable {
     private readonly _onDidChangePackages = new EventEmitter<DidChangePackagesEventArgs>();
@@ -39,7 +28,10 @@ export class CondaPackageManager implements PackageManager, Disposable {
 
     private packages: Map<string, Package[]> = new Map();
 
-    constructor(public readonly api: PythonEnvironmentApi, public readonly log: LogOutputChannel) {
+    constructor(
+        public readonly api: PythonEnvironmentApi,
+        public readonly log: LogOutputChannel,
+    ) {
         this.name = 'conda';
         this.displayName = 'Conda';
         this.description = CondaStrings.condaPackageMgr;
@@ -78,9 +70,8 @@ export class CondaPackageManager implements PackageManager, Disposable {
             },
             async (_progress, token) => {
                 try {
-                    const before = this.packages.get(environment.envId.id) ?? [];
                     const after = await managePackages(environment, manageOptions, this.api, this, token, this.log);
-                    const changes = getChanges(before, after);
+                    const changes = await getPackageChanges(this, environment, after);
                     this.packages.set(environment.envId.id, after);
                     this._onDidChangePackages.fire({ environment: environment, manager: this, changes });
                 } catch (e) {
@@ -104,9 +95,8 @@ export class CondaPackageManager implements PackageManager, Disposable {
                 title: CondaStrings.condaRefreshingPackages,
             },
             async () => {
-                const before = this.packages.get(environment.envId.id) ?? [];
                 const after = await refreshPackages(environment, this.api, this);
-                const changes = getChanges(before, after);
+                const changes = await getPackageChanges(this, environment, after);
                 this.packages.set(environment.envId.id, after);
                 if (changes.length > 0) {
                     this._onDidChangePackages.fire({ environment, manager: this, changes });
